# openapi/tournament-openapi.json
{"openapi":"3.1.0","info":{"title":"Tournament Service API","description":"Tournament lifecycle management API","contact":{"name":"Team Management Backend"},"license":{"name":"Proprietary"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8085","description":"Generated server url"}],"tags":[{"name":"Tournaments","description":"Tournament lifecycle management APIs"}],"paths":{"/api/tournaments/{id}":{"get":{"tags":["Tournaments"],"summary":"Get tournament","description":"Get tournament details by ID","operationId":"getTournament","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tournament not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Tournament state conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Tournaments"],"summary":"Update tournament","description":"Update editable tournament details","operationId":"updateTournament","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTournamentRequest"}}},"required":true},"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tournament not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Tournament state conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/tournaments":{"get":{"tags":["Tournaments"],"summary":"List tournaments","description":"List tournaments with optional season, status and type filters","operationId":"getTournaments","parameters":[{"name":"season","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","description":"Tournament lifecycle status","enum":["ACTIVE","FINISHED","CANCELLED"]}},{"name":"tournamentType","in":"query","required":false,"schema":{"type":"string","description":"Tournament format","enum":["LEAGUE","CUP","FRIENDLY"]}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseTournamentResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseTournamentResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tournament not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Tournament state conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Tournaments"],"summary":"Create tournament","description":"Create a tournament in active status","operationId":"createTournament","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTournamentRequest"}}},"required":true},"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tournament not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Tournament state conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/tournaments/":{"get":{"tags":["Tournaments"],"summary":"List tournaments","description":"List tournaments with optional season, status and type filters","operationId":"getTournaments_1","parameters":[{"name":"season","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","description":"Tournament lifecycle status","enum":["ACTIVE","FINISHED","CANCELLED"]}},{"name":"tournamentType","in":"query","required":false,"schema":{"type":"string","description":"Tournament format","enum":["LEAGUE","CUP","FRIENDLY"]}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseTournamentResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseTournamentResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tournament not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Tournament state conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Tournaments"],"summary":"Create tournament","description":"Create a tournament in active status","operationId":"createTournament_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTournamentRequest"}}},"required":true},"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tournament not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Tournament state conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/tournaments/{id}/finish":{"patch":{"tags":["Tournaments"],"summary":"Finish tournament","description":"Mark an active tournament as finished","operationId":"finishTournament","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tournament not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Tournament state conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/tournaments/{id}/cancel":{"patch":{"tags":["Tournaments"],"summary":"Cancel tournament","description":"Mark an active tournament as cancelled","operationId":"cancelTournament","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tournament not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Tournament state conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"UpdateTournamentRequest":{"type":"object","properties":{"name":{"type":"string","description":"Tournament name","example":"Summer League","maxLength":100,"minLength":0},"tournamentType":{"type":"string","description":"Tournament format","enum":["LEAGUE","CUP","FRIENDLY"],"example":"LEAGUE"},"season":{"type":"string","description":"Season label","example":"2026","maxLength":30,"minLength":0},"startDate":{"type":"string","format":"date","description":"Tournament start date","example":"2026-06-01"},"endDate":{"type":"string","format":"date","description":"Tournament end date","example":"2026-08-31"},"organizer":{"type":"string","description":"Organizer name","example":"City Association","maxLength":100,"minLength":0},"description":{"type":"string","description":"Tournament description"}},"required":["name","season","tournamentType"]},"ErrorResponse":{"type":"object","description":"Tournament API error response","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"},"path":{"type":"string"},"fieldErrors":{"type":"object","additionalProperties":{"type":"string"}}}},"TournamentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"tournamentType":{"type":"string","description":"Tournament format","enum":["LEAGUE","CUP","FRIENDLY"]},"season":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"organizer":{"type":"string"},"description":{"type":"string"},"status":{"type":"string","description":"Tournament lifecycle status","enum":["ACTIVE","FINISHED","CANCELLED"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int64"}}},"CreateTournamentRequest":{"type":"object","properties":{"name":{"type":"string","description":"Tournament name","example":"Summer League","maxLength":100,"minLength":0},"tournamentType":{"type":"string","description":"Tournament format","enum":["LEAGUE","CUP","FRIENDLY"],"example":"LEAGUE"},"season":{"type":"string","description":"Season label","example":"2026","maxLength":30,"minLength":0},"startDate":{"type":"string","format":"date","description":"Tournament start date","example":"2026-06-01"},"endDate":{"type":"string","format":"date","description":"Tournament end date","example":"2026-08-31"},"organizer":{"type":"string","description":"Organizer name","example":"City Association","maxLength":100,"minLength":0},"description":{"type":"string","description":"Tournament description"}},"required":["name","season","tournamentType"]},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageResponseTournamentResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/TournamentResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}